+++ src/main/java/com/floney/floney/analyze/repository/BookAnalyzeRepository.java
@@ -1,0 +1,7 @@
+package com.floney.floney.analyze.repository;
+
+import com.floney.floney.analyze.entity.BookAnalyze;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookAnalyzeRepository extends JpaRepository<BookAnalyze, Long> {
+}
