+++ src/main/java/com/floney/floney/analyze/service/AnalyzeService.java
@@ -1,11 +1,11 @@
 package com.floney.floney.analyze.service;
 
-import com.floney.floney.book.dto.response.AnalyzeResponse;
-import com.floney.floney.book.dto.request.AnalyzeByCategoryRequest;
-import com.floney.floney.book.dto.request.AnalyzeRequestByAsset;
-import com.floney.floney.book.dto.request.AnalyzeRequestByBudget;
-import com.floney.floney.book.dto.response.AnalyzeResponseByAsset;
-import com.floney.floney.book.dto.response.AnalyzeResponseByBudget;
+import com.floney.floney.analyze.dto.response.AnalyzeResponse;
+import com.floney.floney.analyze.dto.request.AnalyzeByCategoryRequest;
+import com.floney.floney.analyze.dto.request.AnalyzeRequestByAsset;
+import com.floney.floney.analyze.dto.request.AnalyzeRequestByBudget;
+import com.floney.floney.analyze.dto.response.AnalyzeResponseByAsset;
+import com.floney.floney.analyze.dto.response.AnalyzeResponseByBudget;
 import com.floney.floney.book.entity.Book;
 
 public interface AnalyzeService {
