--- src/main/java/com/floney/floney/book/repository/BookAnalyzeRepository.java
@@ -1,7 +1,0 @@
-package com.floney.floney.book.repository;
-
-import com.floney.floney.book.entity.BookAnalyze;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BookAnalyzeRepository extends JpaRepository<BookAnalyze, Long> {
-}
