+++ src/main/java/com/floney/floney/book/repository/BookLineCustomRepository.java
@@ -1,12 +1,15 @@
 package com.floney.floney.book.repository;
 
+import com.floney.floney.analyze.dto.request.AnalyzeByCategoryRequest;
+import com.floney.floney.analyze.dto.request.AnalyzeRequestByAsset;
+import com.floney.floney.analyze.dto.request.AnalyzeRequestByBudget;
 import com.floney.floney.book.dto.request.*;
-import com.floney.floney.book.dto.response.AnalyzeResponseByCategory;
+import com.floney.floney.analyze.dto.response.AnalyzeResponseByCategory;
 import com.floney.floney.book.dto.process.BookLineExpense;
 import com.floney.floney.book.dto.process.DayLine;
 import com.floney.floney.book.dto.process.DayLineByDayView;
 import com.floney.floney.book.dto.process.TotalExpense;
-import com.floney.floney.book.dto.response.AnalyzeResponseByBudget;
+import com.floney.floney.analyze.dto.response.AnalyzeResponseByBudget;
 import com.floney.floney.book.entity.BookUser;
 
 import java.time.LocalDate;
